requirements.txt
# Django Framework
Django==5.1.15
djangorestframework==3.15.0

# Database & ORM
psycopg2-binary==2.9.9  # PostgreSQL adapter (optional)

# Image Processing
Pillow==10.1.0

# CORS Support
django-cors-headers==4.3.1

# Environment Variables
python-decouple==3.8

# API Documentation (optional)
drf-spectacular==0.26.5

# Pagination & Filtering (optional)
django-filter==23.5

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code Quality
black==23.12.1
flake8==6.1.0
isort==5.13.2
pylint==3.0.3

# Development
django-extensions==3.2.3
ipython==8.19.0

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0  # Static files serving

# Security
django-environ==0.21.0
